requirements.txt
annotated-types==0.7.0
et_xmlfile==2.0.0
greenlet==3.1.1
mysql-connector-python==9.1.0
numpy==2.1.3
openpyxl==3.1.5
pandas==2.2.3
pydantic==2.10.2
pydantic_core==2.27.1
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.16.0
SQLAlchemy==2.0.36
sqlmodel==0.0.22
typing_extensions==4.12.2
tzdata==2024.2
